chore(root): rename script/utils => script/scripting-utils
