+++ scripts/scripting-utils/colorify.ts
