+++ scripts/scripting-utils/create-scripts.ts
