+++ scripts/scripting-utils/get-all-directories.ts
