+++ scripts/update-package-json.ts
@@ -2,8 +2,8 @@
 import { readdir } from "node:fs/promises";
 import path from "node:path";
 import { $ } from "bun";
-import { colorify } from "./utils/colorify";
-import { createScript, type ScriptConfig, validators } from "./utils/create-scripts";
+import { colorify } from "./scripting-utils/colorify";
+import { createScript, type ScriptConfig, validators } from "./scripting-utils/create-scripts";
 
 const updatePackageJsonConfig = {
 	name: "update-package-json",
